# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Flight Bookings",
        "title": "Flight Bookings",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/flights": {
            "get": {
                "description": "Retrieve a list of all flights",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all flights",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secretKey",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Flight"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new flight based on the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new flight",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secretKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Flight data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateFlightRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/flights/{id}": {
            "get": {
                "description": "Retrieve flight by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get a flight by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secretKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Flight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Flight"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update information about a flight based on the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update a flight",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secretKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Updated flight data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateFlightRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.statusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a flight by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a flight",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key",
                        "name": "secretKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Flight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.statusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "SignIn",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.signInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/sign-up": {
            "post": {
                "description": "Create Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "SignUp",
                "operationId": "create-account",
                "parameters": [
                    {
                        "description": "userID",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "userID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.signInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.statusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Flight": {
            "type": "object",
            "required": [
                "available_business_tickets",
                "available_deluxe_tickets",
                "available_economy_tickets",
                "business_price",
                "deluxe_price",
                "departure_point",
                "destination",
                "economy_price",
                "end_time",
                "start_time",
                "total_business_tickets",
                "total_deluxe_tickets",
                "total_economy_tickets"
            ],
            "properties": {
                "available_business_tickets": {
                    "type": "integer"
                },
                "available_deluxe_tickets": {
                    "type": "integer"
                },
                "available_economy_tickets": {
                    "type": "integer"
                },
                "business_price": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deluxe_price": {
                    "type": "number"
                },
                "departure_point": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "economy_price": {
                    "type": "number"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "total_business_tickets": {
                    "type": "integer"
                },
                "total_deluxe_tickets": {
                    "type": "integer"
                },
                "total_economy_tickets": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "requests.CreateFlightRequest": {
            "type": "object",
            "required": [
                "available_business_tickets",
                "available_deluxe_tickets",
                "available_economy_tickets",
                "business_price",
                "deluxe_price",
                "departure_point",
                "destination",
                "economy_price",
                "end_time",
                "start_time",
                "total_business_tickets",
                "total_deluxe_tickets",
                "total_economy_tickets"
            ],
            "properties": {
                "available_business_tickets": {
                    "type": "integer"
                },
                "available_deluxe_tickets": {
                    "type": "integer"
                },
                "available_economy_tickets": {
                    "type": "integer"
                },
                "business_price": {
                    "type": "number"
                },
                "deluxe_price": {
                    "type": "number"
                },
                "departure_point": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "economy_price": {
                    "type": "number"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "total_business_tickets": {
                    "type": "integer"
                },
                "total_deluxe_tickets": {
                    "type": "integer"
                },
                "total_economy_tickets": {
                    "type": "integer"
                }
            }
        },
        "requests.UpdateFlightRequest": {
            "type": "object",
            "required": [
                "available_business_tickets",
                "available_deluxe_tickets",
                "available_economy_tickets",
                "business_price",
                "deluxe_price",
                "departure_point",
                "destination",
                "economy_price",
                "end_time",
                "start_time",
                "total_business_tickets",
                "total_deluxe_tickets",
                "total_economy_tickets"
            ],
            "properties": {
                "available_business_tickets": {
                    "type": "integer"
                },
                "available_deluxe_tickets": {
                    "type": "integer"
                },
                "available_economy_tickets": {
                    "type": "integer"
                },
                "business_price": {
                    "type": "number"
                },
                "deluxe_price": {
                    "type": "number"
                },
                "departure_point": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "economy_price": {
                    "type": "number"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "total_business_tickets": {
                    "type": "integer"
                },
                "total_deluxe_tickets": {
                    "type": "integer"
                },
                "total_economy_tickets": {
                    "type": "integer"
                }
            }
        },
...
    }
}